azure_vmss_pool/requirements.txt
azure-devops
msrest
pyyaml